+++ src/DependencyInjection/Configuration.php
@@ -28,12 +28,7 @@ class Configuration implements ConfigurationInterface
     {
         $treeBuilder = new TreeBuilder('sonata_notification');
 
-        // Keep compatibility with symfony/config < 4.2
-        if (!method_exists($treeBuilder, 'getRootNode')) {
-            $rootNode = $treeBuilder->root('sonata_notification')->children();
-        } else {
-            $rootNode = $treeBuilder->getRootNode()->children();
-        }
+        $rootNode = $treeBuilder->getRootNode()->children();
 
         $backendInfo = <<<'EOF'
 Other backends you can use:
@@ -192,12 +187,7 @@ EOF;
     {
         $treeBuilder = new TreeBuilder('queues');
 
-        // Keep compatibility with symfony/config < 4.2
-        if (!method_exists($treeBuilder, 'getRootNode')) {
-            $node = $treeBuilder->root('queues');
-        } else {
-            $node = $treeBuilder->getRootNode();
-        }
+        $node = $treeBuilder->getRootNode();
 
         $queuesInfo = <<<'EOF'
 Example for using RabbitMQ