+++ tests/Backend/MessageManagerBackendTest.php
@@ -13,6 +13,9 @@ declare(strict_types=1);
 
 namespace Sonata\NotificationBundle\Tests\Notification;
 
+use Laminas\Diagnostics\Result\Failure;
+use Laminas\Diagnostics\Result\Success;
+use Laminas\Diagnostics\Result\Warning;
 use PHPUnit\Framework\TestCase;
 use Sonata\NotificationBundle\Backend\MessageManagerBackend;
 use Sonata\NotificationBundle\Exception\HandlingException;
@@ -20,9 +23,6 @@ use Sonata\NotificationBundle\Model\MessageInterface;
 use Sonata\NotificationBundle\Model\MessageManagerInterface;
 use Sonata\NotificationBundle\Tests\Entity\Message;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use ZendDiagnostics\Result\Failure;
-use ZendDiagnostics\Result\Success;
-use ZendDiagnostics\Result\Warning;
 
 class MessageManagerBackendTest extends TestCase
 {
@@ -90,12 +90,8 @@ class MessageManagerBackendTest extends TestCase
      */
     public function testStatus($counts, $expectedStatus, $message): void
     {
-        if (!class_exists(Success::class)) {
-            $this->markTestSkipped('The class ZendDiagnostics\Result\Success does not exist');
-        }
-
         $modelManager = $this->createMock(MessageManagerInterface::class);
-        $modelManager->expects($this->exactly(1))->method('countStates')->willReturn($counts);
+        $modelManager->expects($this->once())->method('countStates')->willReturn($counts);
 
         $backend = new MessageManagerBackend($modelManager, [
             MessageInterface::STATE_IN_PROGRESS => 10,
@@ -110,7 +106,7 @@ class MessageManagerBackendTest extends TestCase
         $this->assertSame($message, $status->getMessage());
     }
 
-    public static function statusProvider()
+    public static function statusProvider(): array
     {
         if (!class_exists(Success::class)) {
             return [[1, 1, 1]];