+++ tests/Entity/MessageManagerTest.php
@@ -60,7 +60,7 @@ class MessageManagerTest extends TestCase
         $self = $this;
         $this
             ->getMessageManager(static function ($qb) use ($self): void {
-                $qb->expects($self->once())->method('getRootAliases')->will($self->returnValue(['m']));
+                $qb->expects($self->once())->method('getRootAliases')->willReturn(['m']);
                 $qb->expects($self->never())->method('andWhere');
                 $qb->expects($self->once())->method('setParameters')->with([]);
                 $qb->expects($self->once())->method('orderBy')->with(
@@ -76,9 +76,8 @@ class MessageManagerTest extends TestCase
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Invalid sort field \'invalid\' in \'Sonata\\NotificationBundle\\Entity\\BaseMessage\' class');
 
-        $self = $this;
         $this
-            ->getMessageManager(static function ($qb) use ($self): void {
+            ->getMessageManager(static function ($qb): void {
             })
             ->getPager([], 1, 10, ['invalid' => 'ASC']);
     }
@@ -88,7 +87,7 @@ class MessageManagerTest extends TestCase
         $self = $this;
         $this
             ->getMessageManager(static function ($qb) use ($self): void {
-                $qb->expects($self->once())->method('getRootAliases')->will($self->returnValue(['m']));
+                $qb->expects($self->once())->method('getRootAliases')->willReturn(['m']);
                 $qb->expects($self->never())->method('andWhere');
                 $qb->expects($self->once())->method('setParameters')->with([]);
                 $qb->expects($self->exactly(2))->method('orderBy')->with(
@@ -114,7 +113,7 @@ class MessageManagerTest extends TestCase
         $self = $this;
         $this
             ->getMessageManager(static function ($qb) use ($self): void {
-                $qb->expects($self->once())->method('getRootAliases')->will($self->returnValue(['m']));
+                $qb->expects($self->once())->method('getRootAliases')->willReturn(['m']);
                 $qb->expects($self->once())->method('andWhere')->with($self->equalTo('m.state = :state'));
                 $qb->expects($self->once())->method('setParameters')->with($self->equalTo([
                     'state' => MessageInterface::STATE_OPEN,
@@ -128,7 +127,7 @@ class MessageManagerTest extends TestCase
         $self = $this;
         $this
             ->getMessageManager(static function ($qb) use ($self): void {
-                $qb->expects($self->once())->method('getRootAliases')->will($self->returnValue(['m']));
+                $qb->expects($self->once())->method('getRootAliases')->willReturn(['m']);
                 $qb->expects($self->once())->method('andWhere')->with($self->equalTo('m.state = :state'));
                 $qb->expects($self->once())->method('setParameters')->with($self->equalTo([
                     'state' => MessageInterface::STATE_CANCELLED,
@@ -142,7 +141,7 @@ class MessageManagerTest extends TestCase
         $self = $this;
         $this
             ->getMessageManager(static function ($qb) use ($self): void {
-                $qb->expects($self->once())->method('getRootAliases')->will($self->returnValue(['m']));
+                $qb->expects($self->once())->method('getRootAliases')->willReturn(['m']);
                 $qb->expects($self->once())->method('andWhere')->with($self->equalTo('m.state = :state'));
                 $qb->expects($self->once())->method('setParameters')->with($self->equalTo([
                     'state' => MessageInterface::STATE_IN_PROGRESS,
@@ -166,22 +165,14 @@ class MessageManagerTest extends TestCase
             ->findByTypes([], MessageInterface::STATE_OPEN, 10);
     }
 
-    /**
-     * @return MessageManagerMock
-     */
-    protected function getMessageManagerMock()
+    protected function getMessageManagerMock(): MessageManagerMock
     {
         $registry = $this->createMock(ManagerRegistry::class);
 
-        $manager = new MessageManagerMock(Message::class, $registry);
-
-        return $manager;
+        return new MessageManagerMock(Message::class, $registry);
     }
 
-    /**
-     * @return MessageManager
-     */
-    protected function getMessageManager($qbCallback)
+    protected function getMessageManager($qbCallback): MessageManager
     {
         $query = $this->getMockForAbstractClass(
             AbstractQuery::class,
@@ -192,42 +183,40 @@ class MessageManagerTest extends TestCase
             true,
             ['execute']
         );
-        $query->expects($this->any())->method('execute')->willReturn(true);
+        $query->method('execute')->willReturn(true);
 
         $qb = $this->getMockBuilder(QueryBuilder::class)
             ->setConstructorArgs([$this->createMock(EntityManager::class)])
             ->getMock();
 
-        $qb->expects($this->any())->method('select')->willReturn($qb);
-        $qb->expects($this->any())->method('getQuery')->willReturn($query);
+        $qb->method('select')->willReturn($qb);
+        $qb->method('getQuery')->willReturn($query);
 
         $qbCallback($qb);
 
         $repository = $this->createMock(EntityRepository::class);
-        $repository->expects($this->any())->method('createQueryBuilder')->willReturn($qb);
+        $repository->method('createQueryBuilder')->willReturn($qb);
 
         $metadata = $this->createMock(ClassMetadata::class);
-        $metadata->expects($this->any())->method('getFieldNames')->willReturn([
+        $metadata->method('getFieldNames')->willReturn([
             'state',
             'type',
         ]);
 
         $em = $this->createMock(EntityManager::class);
-        $em->expects($this->any())->method('getRepository')->willReturn($repository);
-        $em->expects($this->any())->method('getClassMetadata')->willReturn($metadata);
+        $em->method('getRepository')->willReturn($repository);
+        $em->method('getClassMetadata')->willReturn($metadata);
 
         $registry = $this->createMock(ManagerRegistry::class);
-        $registry->expects($this->any())->method('getManagerForClass')->willReturn($em);
+        $registry->method('getManagerForClass')->willReturn($em);
 
         return  new MessageManager(BaseMessage::class, $registry);
     }
 
     /**
      * @param int $state
-     *
-     * @return Message
      */
-    protected function getMessage($state = MessageInterface::STATE_OPEN)
+    protected function getMessage($state = MessageInterface::STATE_OPEN): Message
     {
         $message = new Message();